+++ docs/changes.rst
@@ -42,6 +42,9 @@ Released: not yet
 * Test: Circumvented unicode issue with lxml.etree.fromstring()/XML() on
   Python 3.9 by passing in binary strings.
 
+* Test: Adjusted _format()/_ascii2() testcases to PyPy3 behavior with binary vs
+  unicode results.
+
 **Enhancements:**
 
 * Logging: Added a value 'off' for the log destination in the